配仓信息

+++ clms-warehouse-manage/warehouse-api/src/main/java/com/lcdt/warehouse/entity/InorderGoodsInfo.java
@@ -66,11 +66,11 @@ public class InorderGoodsInfo implements Serializable {
     /**
      * 入库价
      */
-    private Integer inHousePrice;
+    private Float inHousePrice;
     /**
      * 货损
      */
-    private Integer damage;
+    private Float damage;
     /**
      * 备注
      */
@@ -220,19 +220,19 @@ public class InorderGoodsInfo implements Serializable {
         this.inHouseAmount = inHouseAmount;
     }
 
-    public Integer getInHousePrice() {
+    public Float getInHousePrice() {
         return inHousePrice;
     }
 
-    public void setInHousePrice(Integer inHousePrice) {
+    public void setInHousePrice(Float inHousePrice) {
         this.inHousePrice = inHousePrice;
     }
 
-    public Integer getDamage() {
+    public Float getDamage() {
         return damage;
     }
 
-    public void setDamage(Integer damage) {
+    public void setDamage(Float damage) {
         this.damage = damage;
     }
